+++ src/Form.fs
@@ -7,8 +7,11 @@ open Fable.React.Props
 
 [<RequireQualifiedAccess>]
 module Form =
-    type FormProps = Custom of HTMLAttr list
+    type FormProps = Custom of IHTMLProp list
 
     let form (props: FormProps seq) (elems: ReactElement seq): ReactElement =
-        let props = keyValueList CaseRules.LowerFirst props
+        let props =
+            props
+            |> Seq.collect (function Custom props -> props)
+            |> keyValueList CaseRules.LowerFirst
         ofImport "Form" "reactstrap" props elems