+++ src/Input.fs
@@ -10,6 +10,35 @@ open Fable.React.Props
 [<RequireQualifiedAccess>]
 module Input =
 
+    [<StringEnum>]
+    type InputType =
+        | [<CompiledName("text")>] Text
+        | [<CompiledName("email")>] Email
+        | [<CompiledName("select")>] Select
+        | [<CompiledName("file")>] File
+        | [<CompiledName("radio")>] Radio
+        | [<CompiledName("checkbox")>] Checkbox
+        | [<CompiledName("textarea")>] Textarea
+        | [<CompiledName("button")>] Button
+        | [<CompiledName("reset")>] Reset
+        | [<CompiledName("submit")>] Submit
+        | [<CompiledName("date")>] Date
+        | [<CompiledName("datetime-local")>] DatetimeLocal
+        | [<CompiledName("hidden")>] Hidden
+        | [<CompiledName("image")>] Image
+        | [<CompiledName("month")>] Month
+        | [<CompiledName("number")>] Number
+        | [<CompiledName("range")>] Range
+        | [<CompiledName("search")>] Search
+        | [<CompiledName("tel")>] Tel
+        | [<CompiledName("url")>] Url
+        | [<CompiledName("week")>] Week
+        | [<CompiledName("password")>] Password
+        | [<CompiledName("datetime")>] Datetime
+        | [<CompiledName("time")>] Time
+        | [<CompiledName("color")>] Color
+
+
     type InputProps =
         | Size of Common.Size
         | BsSize of Common.Size
@@ -20,9 +49,10 @@ module Input =
         | InnerRef of (Element -> unit)
         | CssModule of CSSModule
         | Tag of U2<string, obj>
+        | Type of InputType
         | Custom of IHTMLProp list
 
-    let input (props: InputProps seq) (elems: ReactElement seq): ReactElement =
+    let input (props: InputProps seq) : ReactElement =
         let customProps =
             props
             |> Seq.collect (function
@@ -38,4 +68,4 @@ module Input =
             |> keyValueList CaseRules.LowerFirst
 
         let props = JS.Object.assign (createEmpty, customProps, typeProps)
-        ofImport "Input" "reactstrap" props elems
+        ofImport "Input" "reactstrap" props []