+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 0.3.0
+
+* Corrected props for `Form`
+* Added `Check` to `Label`
+* Added `InputType` enum
+
 ### 0.2.0
 
 * Renamed `ReactStrap` to `Reactstrap`